Solution
For the max case, sacrifice the smallest of the leftmost and rightmost gap from the total amount of gaps. For the min case, use sliding window to find the optimal interval of length n and get the minimum amount. There is a special edge case where n-1 cows are contiguous and 1 extra cow is seperate form them. Handle this case seperately.